Revert "refactor/content-scripts-arc Modularize ParaCard, replace card-manage…"
This reverts commit 2f222ddd1fcc21060ad748a26afd87cf36210863.
