+++ package.json
@@ -30,13 +30,16 @@
     "defuddle": "^0.6.6",
     "es-toolkit": "^1.39.10",
     "lucide-vue-next": "^0.544.0",
-    "openai": "^5.23.2",
+    "motion-v": "^1.7.1",
+    "openai": "^5.23.0",
+    "overtype": "^1.2.6",
     "pinia": "^3.0.3",
     "vue": "3.5.21",
     "vue-codemirror6": "^1.3.22",
     "vue-json-pretty": "^2.5.0",
     "vue-markdown-render": "^2.3.0",
-    "vue-router": "^4.5.1"
+    "vue-router": "^4.5.1",
+    "zod": "^4.1.11"
   },
   "devDependencies": {
     "@fontsource-variable/inter": "^5.2.8",
@@ -44,16 +47,16 @@
     "@prettier/plugin-oxc": "^0.0.4",
     "@tailwindcss/typography": "^0.5.19",
     "@tailwindcss/vite": "^4.1.13",
-    "@types/node": "^24.6.0",
+    "@types/node": "^24.5.2",
     "@wxt-dev/auto-icons": "^1.1.0",
     "@wxt-dev/module-vue": "^1.0.3",
-    "daisyui": "^5.1.25",
+    "daisyui": "^5.1.18",
     "globals": "^16.4.0",
-    "happy-dom": "^19.0.2",
+    "happy-dom": "^18.0.1",
     "husky": "^9.1.7",
     "jsdom": "^27.0.0",
-    "lint-staged": "^16.2.3",
-    "oxlint": "^1.19.0",
+    "lint-staged": "^16.2.0",
+    "oxlint": "^1.18.0",
     "oxlint-tsgolint": "^0.2.0",
     "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.6.14",
@@ -61,7 +64,7 @@
     "typescript": "5.9.2",
     "vite-tsconfig-paths": "^5.1.4",
     "vitest": "^3.2.4",
-    "vue-tsc": "^3.1.0",
+    "vue-tsc": "^3.0.8",
     "wxt": "^0.20.11"
   },
   "overrides": {
