+++ src/agent/agent.ts
@@ -21,6 +21,7 @@ export class LangAgent implements LangAgentSpec {
   private readonly log = createLogger('agent');
   readonly taskTypes = TASK_TYPES;
   private configService = new TaskConfigService();
+  private clientPool = new OpenAIClientPool();
   private readonly runsMutex = new Mutex();
 
   async init() {
@@ -28,22 +29,10 @@ export class LangAgent implements LangAgentSpec {
     await this.configService.init();
   }
 
-  async dispose() {
+  dispose() {
     this.log.info`Disposing LangAgent`;
     this.configService.dispose();
-
-    // Clear the client pool if it was initialized
-    if (clientPoolPromise) {
-      try {
-        const clientPool = await clientPoolPromise;
-        clientPool.clear();
-      } catch (err) {
-        // Ignore errors during cleanup
-        this.log.warn`Failed to clear client pool during dispose: ${err}`;
-      }
-      // Reset the promise so a new pool can be created if needed
-      clientPoolPromise = null;
-    }
+    this.clientPool.clear();
   }
 
   async perform(taskType: TaskType, context: AgentContext): Promise<AgentResponse> {
@@ -51,10 +40,7 @@ export class LangAgent implements LangAgentSpec {
       this.log.info`Performing task: ${taskType}...`;
 
       const config = this.configService.get(taskType);
-
-      // Get OpenAI client pool using single-flight initialization
-      const clientPool = await getClientPool();
-      const client = await clientPool.get(config.aiConfigId);
+      const client = await this.clientPool.get(config.aiConfigId);
       const runner = RUNNERS[taskType];
 
       const response = await runner.run(context, config, client);
@@ -97,7 +83,6 @@ export class LangAgent implements LangAgentSpec {
 }
 
 let langAgentPromise: Promise<LangAgent> | null = null;
-let clientPoolPromise: Promise<OpenAIClientPool> | null = null;
 
 export async function getLangAgent(): Promise<LangAgent> {
   if (!langAgentPromise) {
@@ -113,14 +98,4 @@ export async function getLangAgent(): Promise<LangAgent> {
     })();
   }
   return langAgentPromise;
-}
-
-async function getClientPool(): Promise<OpenAIClientPool> {
-  if (!clientPoolPromise) {
-    clientPoolPromise = (async () => {
-      const { OpenAIClientPool } = await import('@/agent/services/openai-client-pool');
-      return new OpenAIClientPool();
-    })();
-  }
-  return clientPoolPromise;
 }
