+++ src/entrypoints/content/cache-manager.ts
