+++ src/entrypoints/content/content-utils.ts
@@ -1,16 +1,19 @@
-import { updateSharedStyleContent } from '@/entrypoints/content/card/ui/style-manager';
 import { preferenceStorage } from '@/stores/preference';
-import { createLogger } from '@/utils/logger';
 
 import type { Preference } from '@/stores/preference';
 
-const logger = createLogger('content-store');
+import { logger } from './index';
+
+export const isEditable = (element: Element | null): boolean => {
+  if (!(element instanceof HTMLElement)) {
+    return false;
+  }
+  const tagName = element.tagName.toUpperCase();
+  return tagName === 'INPUT' || tagName === 'TEXTAREA' || element.isContentEditable;
+};
 
 export let contentStore: Preference | null = null;
 
-/**
- * Initialize the content store by loading preferences and watching for changes.
- */
 export const initContentStore = () => {
   preferenceStorage
     .getValue()
@@ -26,7 +29,6 @@ export const initContentStore = () => {
   preferenceStorage.watch((value) => {
     logger.debug`content store updated with ${value}`;
     contentStore = value ?? null;
-    updateSharedStyleContent();
     return;
   });
 };
