+++ src/entrypoints/content/event-handler.ts
@@ -2,7 +2,7 @@ import { onKeyDown, useElementByPoint, useMouse } from '@vueuse/core';
 import { throttle } from 'es-toolkit';
 import { effectScope, watch } from 'vue';
 
-import { toggleParaCard } from '@/entrypoints/content/card';
+import { toggleParaCard } from '@/entrypoints/content/card-manager';
 import { findClosestTextContainer } from '@/utils/paragraph';
 
 import type { ContentScriptContext } from '#imports';
