+++ src/stores/preference.ts
@@ -5,7 +5,6 @@ import { ref } from 'vue';
 
 // default css string
 import defaultParaCardCSS from '@/assets/ParaCard.css?raw';
-
 import { createLogger } from '@/utils/logger';
 
 const logger = createLogger('store', 'preference');
