+++ src/utils/__tests__/template.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest';
 
-import { renderTemplate } from '../template';
-
 import type { AgentContext } from '@/agent/types';
+
+import { renderTemplate } from '../template';
 
 describe('renderTemplate', () => {
   const context: AgentContext = {
